add module for neq_assign
